Create THM Pickle Rick
A walkthrough of my process in completing the Try Hack Me CTF room 'Pickle Rick'. Published for educational purposes to demonstrate effective (and some not so effective) strategies in enumerating a vulnerable http server and exfiltrating sensitive data.

This web server exploitation activity was completed entirely in simulated environments with the explicit consent of the target host. 

Techniques demonstrated are not to be applied to entities without their knowledge and consent.