Create THM Anonforce Boot2Root
A walkthrough of my process in completing the Anonforce room on TryHackMe. Published for educational purposes to demonstrate effective strategies in enumerating a target, exploiting vulnerabilities related to authentication, exfiltrating sensitive data and elevating privilege to root acces.

This Boot2Root activity was completed entirely in a simulated environment with the explicit consent of the target. 

Techniques demonstrated are not to be applied to entities without their knowledge and consent.