Create THM Ignite CTF
A walkthrough of my process in completing the Ignite room on TryHackMe. Published for educational purposes to demonstrate effective strategies in enumerating a target, exploiting vulnerabilities on an web server with a published CVE, exfiltrating sensitive data and elevating privilege to root access.

This Boot2Root activity was completed entirely in a simulated environment with the explicit consent of the target. 

Techniques demonstrated are not to be applied to entities without their knowledge and consent.