Create THM Library Boot2Root
A walkthrough of my process in completing the Library room on TryHackMe. Published for educational purposes to demonstrate effective strategies in enumerating a target and exploiting authentication vulnerabilities on a web server to elevate privilege to root and exfiltrate sensitive data

This Boot2Root activity was completed entirely in a simulated environment with the explicit consent of the target. 

Techniques demonstrated are not to be applied to entities without their knowledge and consent.